feature(demo): Update all demos to use synth. sources

+++ apps/ng-sa-counter/src/app/app.component.ts
@@ -32,5 +32,7 @@ export class AppComponent {
     reset: this.resetBoth$,
   });
 
-  constructor(private adapt: AdaptCommon) {}
+  constructor(private adapt: AdaptCommon) {
+    this.store1.double();
+  }
 }
